Income Histories: Added RecyclerView with backend support

+++ app/src/main/java/com/felixfavour/mobotithe/gui/view/income/SubmitIncomeFragment.kt
@@ -36,7 +36,7 @@ class SubmitIncomeFragment : Fragment() {
             // Creation of an IncomeHistory Object
             val incomeHistory = IncomeHistory(
                 Date(binding.calendar.date),
-                binding.amount.text.toString().toDouble(),
+                binding.amount.text.toString().toInt(),
                 income
             )
